refactor: peer type and util references for now
