--- src/chart_types/goal_chart/layout/types/geometry_types.ts
@@ -1,34 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License. */
-
-// In preparation of nominal types in future TS versions
-// https://github.com/microsoft/TypeScript/pull/33038
-// eg. to avoid adding angles and coordinates and similar inconsistent number/number ops.
-// could in theory be three-valued (in,on,out)
-// It also serves as documentation.
-
-export type Pixels = number;
-export type Ratio = number;
-export type SizeRatio = Ratio;
-type Cartesian = number;
-export type Coordinate = Cartesian;
-
-export interface PointObject {
-  x: Coordinate;
-  y: Coordinate;
-}
