--- src/chart_types/goal_chart/layout/utils/measure.ts
@@ -1,32 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License. */
-
-import { Box, Font, TextMeasure } from '../types/types';
-import { Pixels } from '../types/geometry_types';
-
-export function cssFontShorthand({ fontStyle, fontVariant, fontWeight, fontFamily }: Font, fontSize: Pixels) {
-  return `${fontStyle} ${fontVariant} ${fontWeight} ${fontSize}px ${fontFamily}`;
-}
-
-export function measureText(ctx: CanvasRenderingContext2D): TextMeasure {
-  return (fontSize: number, boxes: Box[]): TextMetrics[] =>
-    boxes.map((box: Box) => {
-      ctx.font = cssFontShorthand(box, fontSize);
-      return ctx.measureText(box.text);
-    });
-}
